+++ project/src/components/app/app.tsx
@@ -1,14 +1,14 @@
-import {Route, BrowserRouter, Routes} from 'react-router-dom';
+import {BrowserRouter, Route, Routes} from 'react-router-dom';
 import MainPage from '../../pages/main-page/main-page';
 import MyListPage from '../../pages/my-list/my-list-page';
-import MoviePage from '../../pages/film/movie-page';
-import PlayerPage from '../../pages/player/player-page';
 import SignInPage from '../../pages/sign-in/sign-in-page';
-import AddReviewPage from '../../pages/add-review/add-review-page';
 import NotFoundPage from '../../pages/not-found/not-found-page';
 import PrivateRoute from '../private-route/private-route';
 import {AppRoutes, AuthorizationStatus} from './const';
 import {Films} from '../../types/film';
+import AddReviewOrNotFound from '../../pages/add-review/add-review-or-not-found';
+import MovieOrNotFound from '../../pages/film/film-or-not-found';
+import PlayerOrNotFound from '../../pages/player/player-or-not-found';
 
 type AppProps = {
   promo: {
@@ -30,7 +30,7 @@ function App({promo, films}: AppProps): JSX.Element {
           element={<MainPage promo={promo} films={films}/>}
         >
         </Route>
-        <Route path={AppRoutes.AddReview} element={<AddReviewPage films={films}/>}></Route>
+        <Route path={AppRoutes.AddReview} element={<AddReviewOrNotFound films={films}/>}></Route>
         <Route
           path={AppRoutes.MyList}
           element={
@@ -40,8 +40,8 @@ function App({promo, films}: AppProps): JSX.Element {
           }
         >
         </Route>
-        <Route path={AppRoutes.Film} element={<MoviePage films={films}/>}></Route>
-        <Route path={AppRoutes.Player} element={<PlayerPage films={films}/>}></Route>
+        <Route path={AppRoutes.Film} element={<MovieOrNotFound films={films}/>}></Route>
+        <Route path={AppRoutes.Player} element={<PlayerOrNotFound films={films}/>}></Route>
         <Route path={AppRoutes.SignIn} element={<SignInPage/>}></Route>
         <Route path="*" element={<NotFoundPage/>}></Route>
       </Routes>
